refactor: spawn balloon from runtime scene
