+++ crates/buoy-common/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod objects;
 pub mod plugins;
 pub mod states;
 pub mod types;
