+++ crates/buoy-physics/src/objects/balloon.rs
@@ -1,0 +1,6 @@
+use crate::mesh_utils::{RestState, Strain, create_icosphere_mesh};
+use bevy::prelude::*;
+
+#[derive(Component)]
+#[require(Transform)]
+pub struct Balloon;
