+++ crates/buoy-runtime/src/scene.rs
@@ -5,7 +5,8 @@
 //! config file or spawned at runtime.
 
 use bevy::prelude::*;
-use buoy_common::objects::balloon::spawn_balloon;
+use buoy_physics::objects::balloon::Balloon;
+use buoy_physics::mesh_utils::{create_icosphere_mesh, Strain, RestState};
 
 pub struct SimpleScenePlugin;
 
@@ -15,6 +16,57 @@ impl Plugin for SimpleScenePlugin {
     }
 }
 
-fn setup_scene(mut commands: Commands) {
-    spawn_balloon(&mut commands);
+
+#[cfg(feature = "gui")]
+fn setup_scene(
+    mut commands: Commands,
+    mut meshes: ResMut<Assets<Mesh>>,
+    mut materials: ResMut<Assets<StandardMaterial>>,
+    mut asset_server: Res<AssetServer>,
+) {
+    let balloon_mesh = create_icosphere_mesh(1.0, 0);
+    commands.spawn((
+        Balloon,
+        Transform::from_xyz(0.0, 0.5, 0.0),
+        Strain::default(),
+        RestState::from_mesh(&balloon_mesh),
+        Mesh3d(meshes.add(balloon_mesh)),
+        MeshMaterial3d(materials.add(Color::srgb(0.8, 0.1, 0.2))),
+    ));
+
+    commands.spawn((
+        Mesh3d(meshes.add(Plane3d::default().mesh().size(5.0, 5.0))),
+        MeshMaterial3d(materials.add(Color::srgb(0.3, 0.5, 0.3))),
+        Transform::from_xyz(0.0, 0.0, 0.0),
+    ));
+    commands.spawn((
+        PointLight {
+            intensity: 1500.0,
+            shadows_enabled: true,
+            ..Default::default()
+        },
+        Transform::from_xyz(4.0, 8.0, 4.0),
+    ));
+}
+
+
+#[cfg(not(feature = "gui"))]
+fn setup_scene(
+    mut commands: Commands,
+    mut meshes: ResMut<Assets<Mesh>>,
+    mut asset_server: Res<AssetServer>,
+) {
+    let balloon_mesh = create_icosphere_mesh(1.0, 0);
+    commands.spawn((
+        Balloon,
+        Transform::from_xyz(0.0, 0.5, 0.0),
+        Strain::default(),
+        RestState::from_mesh(&balloon_mesh),
+        Mesh3d(meshes.add(balloon_mesh)),
+    ));
+
+    commands.spawn((
+        Mesh3d(meshes.add(Plane3d::default().mesh().size(5.0, 5.0))),
+        Transform::from_xyz(0.0, 0.0, 0.0),
+    ));
 }
