+++ crates/buoy-ui/Cargo.toml
@@ -7,11 +7,11 @@ authors = { workspace = true }
 license = { workspace = true }
 
 [dependencies]
-buoy-common = { path = "../buoy-common", features = [ "window" ] }
+buoy-common = { path = "../buoy-common", features = ["window"] }
 buoy-physics = { path = "../buoy-physics" }
 bevy = { workspace = true, features = [
     "bevy_asset",
-    "bevy_render", # this includes bevy_mesh
+    "bevy_render",               # this includes bevy_mesh
     "bevy_core_pipeline",
     "bevy_pbr",
     "bevy_text",
@@ -24,12 +24,12 @@ bevy = { workspace = true, features = [
     "tonemapping_luts",
     "x11",
 ] }
-avian3d = { workspace = true, optional = true, features = [ "bevy_picking" ] }
-big_space = { workspace = true, optional = true, features = [ "camera" ] }
+avian3d = { workspace = true, optional = true, features = ["bevy_picking"] }
+big_space = { workspace = true, optional = true, features = ["camera"] }
 bevy_egui = { version = "0.34.1" }
 
 [features]
-default = ["dev", "egui"]
+default = ["dev"]
 dev = [
     "buoy-common/dev",
     "buoy-physics/dev",
@@ -38,7 +38,6 @@ dev = [
     "avian3d/debug-plugin",
     "bevy/bevy_dev_tools",
 ]
-egui = [ "bevy_egui" ]
 grid_space = [
     "buoy-common/grid_space",
     "buoy-physics/grid_space",
